feat(data): Rename Analytics::RevenueStreamsService to DataApi::RevenueStreamsService
