+++ app/services/analytics/base_service.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "lago_http_client"
-
 module Analytics
   class BaseService < BaseService
     def initialize(organization, **filters)
@@ -14,23 +12,5 @@ module Analytics
     private
 
     attr_reader :organization, :filters, :records
-
-    def http_client
-      LagoHttpClient::Client.new(endpoint_url)
-    end
-
-    def headers
-      {
-        "Authorization" => "Bearer #{ENV["LAGO_DATA_API_BEARER_TOKEN"]}"
-      }
-    end
-
-    def endpoint_url
-      "#{ENV["LAGO_DATA_API_URL"]}/#{action_path}"
-    end
-
-    def action_path
-      raise NotImplementedError
-    end
   end
 end
